Added new command and environment to TestflightManager
This commit adds a new command and environment to the TestflightManager package. The new command, PurgeCommand, is used to purge TestFlight users from App Store Connect. The new environment, PurgeEnvironment, is used to configure the PurgeCommand. This change improves the TestflightManager package by adding new functionality for managing TestFlight users.